Implement and use `Test` and `TestModule` types
`shine.run_suite/1` doesn't adhere to the proposed `TestFramework` type,
as it returns a list of test results instead of `Nil`. This wasn't
changed because it would hurt testability. A wrapper function that does
not return the test results will be implemented when required.

Also, the `Test` type doesn't include `name` or `sync` fields, as
they're not required by Shine yet. The former will be implemented when
it gains more elaborate output, and the latter when it gains the ability
to run tests asynchronously.

Closes #6.